+++ versions.tf
@@ -3,12 +3,16 @@ terraform {
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = "~> 3.39"
+      version = "~> 3.67"
     }
     # tflint-ignore: terraform_unused_required_providers
     azurecaf = {
       source  = "aztfmod/azurecaf"
       version = "~> 1.2, >= 1.2.22"
+    }
+    tls = {
+      source  = "hashicorp/tls"
+      version = ">= 3.0"
     }
   }
 }