+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Unreleased
+
+Breaking
+  * AZ-1207: Rework module
+  * AZ-1207: Use directly linux-vm module
+  * AZ-1207: Use Claranet golden image by default
+  * AZ-1207: Rework variables
+
 # v7.5.0 - 2023-07-21
 
 Added